+++ MTP/interface/locations/AddPhotoVC.swift
@@ -56,7 +56,7 @@ final class AddPhotoVC: UIViewController {
         }
     }
 
-    /// Prepare for interaction
+    /// :nodoc:
     override func viewDidLoad() {
         super.viewDidLoad()
         requireOutlets()
@@ -73,9 +73,7 @@ final class AddPhotoVC: UIViewController {
         stopKeyboardListening()
     }
 
-    /// Prepare for reveal
-    ///
-    /// - Parameter animated: Whether animating
+    /// :nodoc:
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
 
@@ -83,9 +81,7 @@ final class AddPhotoVC: UIViewController {
         expose()
     }
 
-    /// Actions to take after reveal
-    ///
-    /// - Parameter animated: Whether animating
+    /// :nodoc:
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         report(screen: "Add Photo")
@@ -107,11 +103,7 @@ final class AddPhotoVC: UIViewController {
         view.endEditing(true)
     }
 
-    /// Instrument and inject navigation
-    ///
-    /// - Parameters:
-    ///   - segue: Navigation action
-    ///   - sender: Action originator
+    /// :nodoc:
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         if let target = Segues.showCountry(segue: segue)?
                               .destination