+++ MTP/interface/profile/FaqVC.swift
@@ -84,7 +84,7 @@ final class FaqVC: UITableViewController {
                      isExpanded: false)
     ]
 
-    /// Prepare for interaction
+    /// :nodoc:
     override func viewDidLoad() {
         super.viewDidLoad()
         requireOutlets()
@@ -96,9 +96,7 @@ final class FaqVC: UITableViewController {
         tableView.rowHeight = UITableView.automaticDimension
     }
 
-    /// Prepare for reveal
-    ///
-    /// - Parameter animated: Whether animating
+    /// :nodoc:
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
 
@@ -106,9 +104,7 @@ final class FaqVC: UITableViewController {
         expose()
     }
 
-    /// Actions to take after reveal
-    ///
-    /// - Parameter animated: Whether animating
+    /// :nodoc:
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         report(screen: "FAQ")
@@ -254,7 +250,7 @@ final class FaqCell: UITableViewCell {
 
 private extension FaqCell {
 
-    @IBAction func toggleTapped(_ sender: UIButton) {
+    @IBAction func toggleTapped(_ sender: UITapGestureRecognizer) {
         guard let button = toggleButton else { return }
 
         button.isSelected.toggle()