+++ MTP/interface/rankings/RankingsVC.swift
@@ -42,7 +42,7 @@ final class RankingsVC: UIViewController {
     private var searchResults: [String: [SearchResultItemJSON]] = [:]
     private var searchKey: String = ""
 
-    /// Prepare for interaction
+    /// :nodoc:
     override func viewDidLoad() {
         super.viewDidLoad()
         requireOutlets()
@@ -51,9 +51,7 @@ final class RankingsVC: UIViewController {
         configureSearchBar()
     }
 
-    /// Prepare for reveal
-    ///
-    /// - Parameter animated: Whether animating
+    /// :nodoc:
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
 
@@ -61,19 +59,13 @@ final class RankingsVC: UIViewController {
         expose()
     }
 
-    /// Actions to take after reveal
-    ///
-    /// - Parameter animated: Whether animating
+    /// :nodoc:
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         report(screen: "Rankings")
     }
 
-    /// Instrument and inject navigation
-    ///
-    /// - Parameters:
-    ///   - segue: Navigation action
-    ///   - sender: Action originator
+    /// :nodoc:
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         if let counts = Segues.showUserCounts(segue: segue)?
                               .destination,