+++ MTP/support/WKWebViewController.swift
@@ -326,7 +326,7 @@ class WKWebViewController: UIViewController {
         }
     }
 
-    /// Prepare for interaction
+    /// :nodoc:
     override func viewDidLoad() {
         super.viewDidLoad()
 
@@ -370,9 +370,7 @@ class WKWebViewController: UIViewController {
         }
     }
 
-    /// Prepare for reveal
-    ///
-    /// - Parameter animated: Whether animating
+    /// :nodoc:
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)