+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- Expandable visited map viewable when tapped
+
 ## [Version 1.2.0, Build 644] - 2019-10-01
 
 ### Added